TwoStageTransfer/requirements.txt
gin==0.1.6
gin-config==0.2.1
javabridge==1.0.18
joblib==0.14.0
numpy==1.17.3
python-weka-wrapper3==0.1.7
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
